# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PR4 Express API
  version: 1.0.0
  description: |
    Учебный API для работы с товарами. Данные хранятся в JSON (без БД).

servers:
  - url: http://localhost:3000

paths:
  /:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: Сервер жив
          content:
            text/plain:
              schema:
                type: string
              example: Express API is running. Try /api/products

  /api/products:
    get:
      summary: Получить список товаров
      responses:
        "200":
          description: Массив товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      summary: Создать товар
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Созданный товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Некорректный ввод (например, title пустой)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                titleRequired:
                  value: { "error": "title is required" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value: { "error": "Internal Server Error" }

  /api/products/{id}:
    patch:
      summary: Частично обновить товар
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID товара (строка)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Обновлённый товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value: { "error": "Not found" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Удалить товар
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID товара (строка)
      responses:
        "200":
          description: Успешное удаление
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
              example: { "ok": true }
        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value: { "error": "Not found" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: ab12cd34
        title:
          type: string
          example: Яблоки Гала
        category:
          type: string
          example: Фрукты
        description:
          type: string
          example: Сладкие свежие яблоки, 1 кг.
        price:
          type: number
          example: 149
        stock:
          type: number
          example: 35
        rating:
          type: number
          example: 4.7
        imageUrl:
          type: string
          example: https://images.example.com/apple.jpg
      required: [id, title, category, description, price, stock]

    ProductCreate:
      type: object
      properties:
        title:
          type: string
          example: Яблоки Гала
        category:
          type: string
          example: Фрукты
        description:
          type: string
          example: Сладкие свежие яблоки, 1 кг.
        price:
          oneOf:
            - type: number
            - type: string
          example: 149
        stock:
          oneOf:
            - type: number
            - type: string
          example: 35
        rating:
          oneOf:
            - type: number
            - type: string
          example: 4.7
        imageUrl:
          type: string
          example: https://images.example.com/apple.jpg
      required: [title, category, description, price, stock]

    ProductPatch:
      type: object
      description: Частичное обновление (можно передать любые поля, которые поддерживает store)
      additionalProperties: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]